apt-requirements.txt
vim
gpac
fbi
python3-dialog
sqlite3
ssh-pass
